pyproject.toml
[project]
name = "whook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "redis>=4.0.0",
    "rq>=1.10.0",
    "python-dotenv>=0.19.0",
    "requests>=2.32.0",
    "restrictedpython>=8.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "authlib>=1.3.0",
    "itsdangerous>=2.1.2",
    "httpx>=0.27.0",
]
